+++ source/src/sound.h
@@ -36,8 +36,8 @@ uint8_t isMusicOn();
 uint8_t isSoundOn();
 
 #ifdef NINTENDO
-void playNintendoCh1(int REG_10, int REG_11, int REG_12, int REG_13, int REG_14);
-void playNintendoCh2(int REG_21, int REG_22, int REG_23, int REG_24);
+void playNintendoCh1(uint8_t REG_10, uint8_t REG_11, uint8_t REG_12, uint8_t REG_13, uint8_t REG_14);
+void playNintendoCh2(uint8_t REG_21, uint8_t REG_22, uint8_t REG_23, uint8_t REG_24);
 #endif
 
 #ifdef SEGA
